+++ st2client/st2client/client.py
@@ -23,6 +23,7 @@ from st2client.models.core import ResourceManager
 from st2client.models.core import ActionAliasResourceManager
 from st2client.models.core import ActionAliasExecutionManager
 from st2client.models.core import LiveActionResourceManager
+from st2client.models.core import InquiryResourceManager
 from st2client.models.core import TriggerInstanceResourceManager
 from st2client.models.core import PackResourceManager
 from st2client.models.core import ConfigManager
@@ -53,6 +54,8 @@ class Client(object):
             self.endpoints['base'] = os.environ.get('ST2_BASE_URL', DEFAULT_BASE_URL)
 
         api_version = api_version or os.environ.get('ST2_API_VERSION', DEFAULT_API_VERSION)
+
+        self.endpoints['exp'] = '%s:%s/%s' % (self.endpoints['base'], DEFAULT_API_PORT, 'exp')
 
         if api_url:
             self.endpoints['api'] = api_url
@@ -122,6 +125,8 @@ class Client(object):
             models.ConfigSchema, self.endpoints['api'], cacert=self.cacert, debug=self.debug)
         self.managers['LiveAction'] = LiveActionResourceManager(
             models.LiveAction, self.endpoints['api'], cacert=self.cacert, debug=self.debug)
+        self.managers['Inquiry'] = InquiryResourceManager(
+            models.Inquiry, self.endpoints['exp'], cacert=self.cacert, debug=self.debug)
         self.managers['Pack'] = PackResourceManager(
             models.Pack, self.endpoints['api'], cacert=self.cacert, debug=self.debug)
         self.managers['Policy'] = ResourceManager(
@@ -172,6 +177,10 @@ class Client(object):
     @property
     def liveactions(self):
         return self.managers['LiveAction']
+
+    @property
+    def inquiries(self):
+        return self.managers['Inquiry']
 
     @property
     def packs(self):