+++ st2client/st2client/models/core.py
@@ -439,6 +439,10 @@ class LiveActionResourceManager(ResourceManager):
         return self.resource.deserialize(response.json())
 
 
+class InquiryResourceManager(ResourceManager):
+    pass
+
+
 class TriggerInstanceResourceManager(ResourceManager):
     @add_auth_token_to_kwargs_from_env
     def re_emit(self, trigger_instance_id, **kwargs):